(#27) SRP 분리, 관심사 분리, 트랜잭션 최적화
- UserPersistenceService에서 ActivityLogOrchestrator 분리
- BadgeService에서 TierGradientProvider, BadgeFormatter, SvgBadgeRenderer 분리
- GitHub API 호출 시 트랜잭션 일시 중단으로 청크 트랜잭션 유지 시간 최소화
